fix: 레이아웃 추가, 디자인 수정, 불필요한 코드 삭제, 리팩토링
