+++ src/app/(main)/(owner)/[id]/pos/page.tsx
@@ -1,11 +1,3 @@
-"use client";
-
-import SectionHeader from "@/components/SectionHeader";
-
 export default function Pos() {
-  return (
-    <div className="h-full w-full">
-      <SectionHeader title="POS" />
-    </div>
-  );
+  return <div className="h-full w-full">POS</div>;
 }
