+++ src/app/(main)/(owner)/[id]/settings/layout.tsx
@@ -1,0 +1,14 @@
+import PageTitle from "@/app/(main)/_components/PageTitle";
+
+export default async function Layout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      <PageTitle title="설정" />
+      {children}
+    </>
+  );
+}
