+++ src/app/(main)/(owner)/[id]/settings/page.tsx
@@ -1,11 +1,3 @@
-"use client";
-
-import SectionHeader from "@/components/SectionHeader";
-
 export default function Settings() {
-  return (
-    <div className="h-full w-full">
-      <SectionHeader title="설정" />
-    </div>
-  );
+  return <div className="h-full w-full">Settings</div>;
 }
