+++ src/app/(main)/(owner)/[id]/subscription/page.tsx
@@ -1,11 +1,3 @@
-"use client";
-
-import SectionHeader from "@/components/SectionHeader";
-
 export default function Subscribe() {
-  return (
-    <div className="h-full w-full">
-      <SectionHeader title="구독 설정" />
-    </div>
-  );
+  return <div className="h-full w-full">구독 설정</div>;
 }
