+++ src/app/(main)/admin/approvals/layout.tsx
@@ -1,0 +1,10 @@
+import PageTitle from "@/app/(main)/_components/PageTitle";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <PageTitle title="매장 등록 신청 현황" />
+      {children}
+    </>
+  );
+}
