+++ src/app/(main)/admin/users/layout.tsx
@@ -1,0 +1,10 @@
+import PageTitle from "@/app/(main)/_components/PageTitle";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <PageTitle title="회원 관리" />
+      {children}
+    </>
+  );
+}
