+++ src/app/(main)/layout.tsx
@@ -8,7 +8,7 @@ export default function Layout({ children }: { children: React.ReactNode }) {
       <div className="flex h-screen w-screen flex-col overflow-hidden bg-white md:flex-row md:bg-[#F5F5F5]">
         <NewSidebar />
         <MobileHeader />
-        <main className="flex flex-1 flex-col overflow-y-auto rounded-[20px] md:mx-8 md:my-8 md:bg-white md:p-5 md:px-6 md:py-5 lg:p-8 lg:py-8">
+        <main className="flex flex-1 flex-col overflow-y-auto rounded-[20px] md:mx-8 md:my-8 md:bg-white md:px-6 md:py-5 lg:px-8 lg:pt-0 lg:pb-8">
           {children}
         </main>
       </div>
