+++ src/app/stores/page.tsx
@@ -6,7 +6,6 @@ import { Plus } from "lucide-react";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import ResponsiveButton from "@/components/common/ResponsiveButton";
-import SectionHeader from "@/components/SectionHeader";
 import {
   MobileTable,
   MobileTableCell,
@@ -24,9 +23,9 @@ import useOverlay from "@/hooks/use-overlay";
 import { registerStateTranslate } from "@/constants/translates";
 import transformDate from "@/lib/formatting/transformDate";
 import QueryProviders from "@/app/query-providers";
-import StoreApplicationModal from "../store/_components/modals/StoreApplicationModal";
-import PendingAcceptModal from "../store/_components/modals/PendingAcceptModal";
-import useStores from "../store/_hooks/useStores";
+import useStores from "../(main)/(owner)/[id]/store/_hooks/useStores";
+import StoreApplicationModal from "../(main)/(owner)/[id]/store/_components/modals/StoreApplicationModal";
+import PendingAcceptModal from "../(main)/(owner)/[id]/store/_components/modals/PendingAcceptModal";
 
 const itemWidths = {
   "No.": {
@@ -53,6 +52,7 @@ const itemWidths = {
 
 export default function StoreList() {
   const navigate = useRouter();
+
   const [currentPage, setCurrentPage] = useState(1);
 
   const { registrationListQuery } = useStores();
@@ -85,8 +85,7 @@ export default function StoreList() {
   }, [currentPage]);
 
   return (
-    <div className="h-full max-h-screen w-full overflow-y-scroll">
-      <SectionHeader title="매장 등록 신청 현황" />
+    <div className="w-full">
       <Table className="z-10 mt-[-10px] flex w-full flex-col md:mt-4">
         <TableHeader className="w-full">
           <TableRow isHead>
@@ -201,7 +200,7 @@ export default function StoreList() {
         <ResponsiveButton
           variant="outline"
           color="outline-primary"
-          onClick={() => navigate.push("/store/create")}
+          onClick={() => navigate.push("/create")}
           responsiveButtons={{
             lg: { buttonSize: "lg" },
             md: { buttonSize: "sm" },
